+++ src/Application/View/Networking/ChatSessionView.php
@@ -10,6 +10,8 @@ class ChatSessionView
     /** @var User */
     public $otherUser;
 
+    public string $avatarUrl;
+
     /** @var DateTimeInterface */
     public $latestMessageDate;
 
@@ -21,11 +23,13 @@ class ChatSessionView
 
     public function __construct(
         User $otherUser,
+        string $avatarUrl,
         DateTimeInterface $latestMessageDate,
         int $messagesCount,
         int $newMessagesCount
     ) {
         $this->otherUser = $otherUser;
+        $this->avatarUrl = $avatarUrl;
         $this->latestMessageDate = $latestMessageDate;
         $this->messagesCount = $messagesCount;
         $this->newMessagesCount = $newMessagesCount;