+++ tests/Application/Query/Networking/NetworkingQueryHandlerTest.php
@@ -4,6 +4,7 @@ namespace Proximum\Vimeet\Tests\Application\Query\Networking;
 
 use DateTime;
 use PHPUnit\Framework\TestCase;
+use Prophecy\Prophecy\ObjectProphecy;
 use Proximum\Vimeet\Domain\Model\User;
 use Proximum\Vimeet\Domain\Model\Event;
 use Proximum\Vimeet\Domain\Model\Sheet;
@@ -15,12 +16,22 @@ use Proximum\Vimeet\Application\Adapter\NotificationSubscriberInterface;
 use Proximum\Vimeet\Application\Query\Networking\NetworkingQueryHandler;
 use Proximum\Vimeet\Infrastructure\Adapter\Mercure\AbstractNotification;
 use Proximum\Vimeet\Application\Adapter\NotificationSubscriptionsInterface;
+use Proximum\Vimeet\Application\Adapter\RouterInterface;
 use Proximum\Vimeet\Application\View\Networking\ChatSessionView;
 use Proximum\Vimeet\Application\View\Networking\NetworkingView;
+use Proximum\Vimeet\Domain\Model\User\Account;
 
 class NetworkingQueryHandlerTest extends TestCase
 {
-    private $user, $participant, $event, $sheet, $notificationSubescriber, $notificationSubscribtions, $chatSessionRepository, $chatMessageRepository;
+    private ObjectProphecy $user;
+    private ObjectProphecy $participant;
+    private ObjectProphecy $event;
+    private ObjectProphecy $sheet;
+    private ObjectProphecy $notificationSubescriber;
+    private ObjectProphecy $notificationSubscribtions;
+    private ObjectProphecy $chatSessionRepository;
+    private ObjectProphecy $chatMessageRepository;
+    private ObjectProphecy $routerAdapter;
 
     public function setUp(): void
     {
@@ -33,15 +44,16 @@ class NetworkingQueryHandlerTest extends TestCase
         $this->notificationSubscribtions  = $this->prophesize(NotificationSubscriptionsInterface::class);
         $this->chatSessionRepository = $this->prophesize(ChatSessionRepositoryInterface::class);
         $this->chatMessageRepository = $this->prophesize(ChatMessageRepositoryInterface::class);
+        $this->routerAdapter = $this->prophesize(RouterInterface::class);
     }
 
     public function testShouldReturnAViewWithoutSessions(): void
     {
 
         $eventId = 555;
-        $topic = "faketopic";
-        $providerUrl = "http://tourte.com";
-        $fakeSubscriberKey = "a fake subscriber key";
+        $topic = 'faketopic';
+        $providerUrl = 'http://tourte.com';
+        $fakeSubscriberKey = 'a fake subscriber key';
         $now = new DateTime();
         $messageCount = 333;
         $privateChatNewMessages = 0;
@@ -72,7 +84,13 @@ class NetworkingQueryHandlerTest extends TestCase
 
         $query = new NetworkingQuery($this->sheet->reveal(), $this->user->reveal());
 
-        $handler = new NetworkingQueryHandler($this->notificationSubescriber->reveal(), $this->notificationSubscribtions->reveal(), $this->chatMessageRepository->reveal(), $this->chatSessionRepository->reveal());
+        $handler = new NetworkingQueryHandler(
+            $this->notificationSubescriber->reveal(),
+            $this->notificationSubscribtions->reveal(),
+            $this->chatMessageRepository->reveal(),
+            $this->chatSessionRepository->reveal(),
+            $this->routerAdapter->reveal()
+        );
 
         $result =  $handler->handle($query);
 
@@ -96,9 +114,9 @@ class NetworkingQueryHandlerTest extends TestCase
         $otherUser = $this->prophesize(User::class);
 
         $eventId = 555;
-        $topic = "faketopic";
-        $providerUrl = "http://tourte.com";
-        $fakeSubscriberKey = "a fake subscriber key";
+        $topic = 'faketopic';
+        $providerUrl = 'http://tourte.com';
+        $fakeSubscriberKey = 'a fake subscriber key';
         $now = new DateTime();
         $messageCount = 333;
 
@@ -110,16 +128,23 @@ class NetworkingQueryHandlerTest extends TestCase
         $totalMessageCount = 855;
         $privateChatNewMessages = $newMessageCount;
 
+        $account = $this->prophesize(Account::class);
+        $account->getCompleteName()->willReturn('John Doe');
+        $otherUser->getAvatar()->willReturn(null);
+        $otherUser->getAccount()->willReturn($account->reveal());
+
         $sessionsByEventAndUser = [[
-            "otherUser" =>  $otherUser->reveal(),
-            "latestMessageDate" => $now,
-            "messagesCount" =>  $totalMessageCount,
-            "unreadMessages" => [
-                $userId =>   $newMessageCount
+            'otherUser' =>  $otherUser->reveal(),
+            'latestMessageDate' => $now,
+            'messagesCount' =>  $totalMessageCount,
+            'unreadMessages' => [
+                $userId => $newMessageCount
             ]
         ]];
 
-        $privateChatSessions = [new ChatSessionView($otherUser->reveal(), $now, $totalMessageCount, $newMessageCount)];
+        $privateChatSessions = [
+            new ChatSessionView($otherUser->reveal(), '/path/to/initals.jpg', $now, $totalMessageCount, $newMessageCount)
+        ];
 
 
         $this->user->getId()->shouldBeCalled()->willReturn($userId);
@@ -142,9 +167,17 @@ class NetworkingQueryHandlerTest extends TestCase
         $this->chatSessionRepository->findSessionsByEventAndUser($this->event->reveal(), $this->user->reveal())->shouldBeCalled()->willReturn($sessionsByEventAndUser);
         $this->chatMessageRepository->getMessagesCountByLinkableObject($this->event->reveal(), $now)->shouldBeCalled()->willReturn($messageCount);
 
+        $this->routerAdapter->generate('event_chat_avatar', ['name' => 'John Doe'])->willReturn('/path/to/initals.jpg');
+
         $query = new NetworkingQuery($this->sheet->reveal(), $this->user->reveal());
 
-        $handler = new NetworkingQueryHandler($this->notificationSubescriber->reveal(), $this->notificationSubscribtions->reveal(), $this->chatMessageRepository->reveal(), $this->chatSessionRepository->reveal());
+        $handler = new NetworkingQueryHandler(
+            $this->notificationSubescriber->reveal(),
+            $this->notificationSubscribtions->reveal(),
+            $this->chatMessageRepository->reveal(),
+            $this->chatSessionRepository->reveal(),
+            $this->routerAdapter->reveal()
+        );
 
         $result =  $handler->handle($query);
 
@@ -168,9 +201,9 @@ class NetworkingQueryHandlerTest extends TestCase
         $otherUser = $this->prophesize(User::class);
 
         $eventId = 555;
-        $topic = "faketopic";
-        $providerUrl = "http://tourte.com";
-        $fakeSubscriberKey = "a fake subscriber key";
+        $topic = 'faketopic';
+        $providerUrl = 'http://tourte.com';
+        $fakeSubscriberKey = 'a fake subscriber key';
         $now = new DateTime();
         $messageCount = 333;
 
@@ -183,13 +216,15 @@ class NetworkingQueryHandlerTest extends TestCase
         $totalMessageCount = 855;
         $privateChatNewMessages = 0;
 
+        $otherUser->getAvatar()->willReturn('/path/to/avatar.jpg');
+
         // This Sesssion has no messages for user with id 1615 (our User)
 
         $sessionsByEventAndUser = [[
-            "otherUser" =>  $otherUser->reveal(),
-            "latestMessageDate" => $now,
-            "messagesCount" =>  $totalMessageCount,
-            "unreadMessages" => [
+            'otherUser' =>  $otherUser->reveal(),
+            'latestMessageDate' => $now,
+            'messagesCount' =>  $totalMessageCount,
+            'unreadMessages' => [
                 $otherUserId =>   $newMessageCount
             ]
         ]];
@@ -197,8 +232,7 @@ class NetworkingQueryHandlerTest extends TestCase
 
 
 
-        $privateChatSessions = [new ChatSessionView($otherUser->reveal(), $now, $totalMessageCount, 0)];
-
+        $privateChatSessions = [new ChatSessionView($otherUser->reveal(), '/path/to/resized-avatar.jpg', $now, $totalMessageCount, 0)];
 
         $this->user->getId()->shouldBeCalled()->willReturn($userId);
 
@@ -219,9 +253,17 @@ class NetworkingQueryHandlerTest extends TestCase
         $this->chatSessionRepository->findSessionsByEventAndUser($this->event->reveal(), $this->user->reveal())->shouldBeCalled()->willReturn($sessionsByEventAndUser);
         $this->chatMessageRepository->getMessagesCountByLinkableObject($this->event->reveal(), $now)->shouldBeCalled()->willReturn($messageCount);
 
+        $this->routerAdapter->generate('liip_imagine_filter', ['path' => '/path/to/avatar.jpg', 'filter' => 'user_icon'])->willReturn('/path/to/resized-avatar.jpg');
+
         $query = new NetworkingQuery($this->sheet->reveal(), $this->user->reveal());
 
-        $handler = new NetworkingQueryHandler($this->notificationSubescriber->reveal(), $this->notificationSubscribtions->reveal(), $this->chatMessageRepository->reveal(), $this->chatSessionRepository->reveal());
+        $handler = new NetworkingQueryHandler(
+            $this->notificationSubescriber->reveal(),
+            $this->notificationSubscribtions->reveal(),
+            $this->chatMessageRepository->reveal(),
+            $this->chatSessionRepository->reveal(),
+            $this->routerAdapter->reveal()
+        );
 
         $result =  $handler->handle($query);