+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - Hotfix - Improve exception info if meeting not found after evaluation timeout
+- MV-318 - Conference order alphabetical
 
 ## [2.94.1] - 2021-06-03
 ### Fixed